% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_boot_t.R
\name{get.t.stat}
\alias{get.t.stat}
\title{Get a t-statistic}
\usage{
get.t.stat(x, y, var.equal = FALSE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{var.equal}{boolean indicating whether or not to assume equal variance. Defaults to FALSE}
}
\value{
the t-statistic based on datasets x and y
}
\description{
Get a t-statistic
}
\examples{
get.t.stat(rnorm(20, 0, 1), rnorm(30, 0, 1))
get.t.stat(rnorm(20, 5, 1), rnorm(30, 0, 2))
}
