% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbtt.R
\name{rbtt}
\alias{rbtt}
\title{Perform robust bootstrapped t-tests}
\usage{
rbtt(x, y, n.boot, n.cores = 1, method = "combined", conf.level = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{n.boot}{number of bootstrap resamples to perform}

\item{n.cores}{number of cores to use for parallelization. Defaults to 1. If using Windows, set n.cores = 1.}

\item{method}{Which robust bootstrapped t-test to perform. Set `method=1’ for a two-sample t-test under the equal variance assumption, 'method = 2' for a two-sample t-test without the equal variance assumption, and 'method = "both"' to perform both methods simultaneously.}

\item{conf.level}{Desired confidence level for computing confidence intervals: a number between 0 and 1.}
}
\value{
p-value of the test
}
\description{
Perform robust bootstrapped t-tests that aim to better control type-I error rates
when comparing means of non-negative distributions with excess zero observations.
}
\examples{
x=rbinom(50,1,0.5)*rlnorm(50,0,1)
y=rbinom(50,1,0.3)*rlnorm(150,2,1)
rbtt(x,y,n.boot=999)
rbtt(x,y,n.boot=9999,n.cores=2)
}
