% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_boot_t.R
\name{robust.boot.t.1}
\alias{robust.boot.t.1}
\title{Perform robust bootstrapped t-test 1}
\usage{
robust.boot.t.1(x, y, n.boot, n.cores = 1)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{n.boot}{number of bootstrap resamples to perform}

\item{n.cores}{number of cores to use for parallelization. Defaults to 1.}
}
\value{
p-value of the test
}
\description{
Perform robust bootstrapped t-test 1
}
\examples{
robust.boot.t.1(rnorm(20, 0, 1), rnorm(30, 0, 1), n.boot = 999)
robust.boot.t.1(rnorm(20, 0, 1), rnorm(30, 5, 1), n.boot = 999, n.cores = 2)
}
