% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_boot_t.R
\name{robust.boot.t}
\alias{robust.boot.t}
\title{Perform robust bootstrapped t-tests}
\usage{
robust.boot.t(x, y, n.boot, n.cores = 1, method = "combined",
  conf.level = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{n.boot}{number of bootstrap resamples to perform}

\item{n.cores}{number of cores to use for parallelization. Defaults to 1. If using Windows, set n.cores = 1.}

\item{method}{Which robust bootstrapped t-test to perform. Set 'method = 1' for method 1, 'method = 1' for method 2, and 'method = "both"' to perform methods 1 and 2 simultaneously.}
}
\value{
p-value of the test
}
\description{
Perform robust bootstrapped t-tests
}
\examples{
robust.boot.t(rnorm(20, 0, 1), rnorm(30, 1, 1), n.boot = 999)
robust.boot.t(rnorm(20, 0, 1), rnorm(30, 5, 1), n.boot = 9999, n.cores = 2)
}
