% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_boot_t.R
\name{robust.boot.t}
\alias{robust.boot.t}
\title{Perform a robust bootstrapped t-test}
\usage{
robust.boot.t(x, y, n.boot, n.cores = 1, algorithm = "combined")
}
\arguments{
\item{x}{a vector of numeric data points}

\item{y}{a vector of numeric data points}

\item{n.boot}{number of bootstrap resamples to perform}

\item{n.cores}{number of cores to use for parallelization. Defaults to 1. If using Windows, set n.cores = 1.}

\item{algorithm}{Which robust bootstrapped t-test to perform. Set 'algorithm = 1' for method 1, 'algorithm = 1' for method 2, and 'algorithm = "both"' to perform methods 1 and 2 simultaneously.}
}
\value{
p-value of the test
}
\description{
Perform a robust bootstrapped t-test
}
\examples{
robust.boot.t(rnorm(20, 0, 1), rnorm(30, 0, 1), n.boot = 999)
robust.boot.t(rnorm(20, 0, 1), rnorm(30, 5, 1), n.boot = 999, n.cores = 2)
}
